perf: reduce gc time spent
This refactors the periodic stats collection and recording so that the provider records count is run less frequently than the others. This also runs each task in it's own goroutine so that errors in one task do not effect the others.

Provider records count is now run every minute (was every 5 seconds) which should reduce the amount of garbage to collect overall as well as the time spent doing it.

refs https://github.com/libp2p/hydra-booster/pull/53