feat: leveldb peerstore (#105)
This PR adds an EXPERIMENTAL leveldb peerstore to the hydra. It can be enabled by passing the `-pstore` flag or by setting an environment variable `HYDRA_PSTORE`. If enabling the leveldb peerstore the value the flag/env var should have is the path on disk where the database should be created/opened.