Add DynamoDB provider store
This adds support for storing provider records in DynamoDB. The reason
for adding this support is that the access patterns required for
provider stores fit key-value stores nicely, which are easier to scale
and have better performance.

This also changes the name of the existing "find providers" metrics,
since that name ended up being confusing, instead we have started
calling the part of the hydras that proactively lookup providers as
"provider prefetching".