Use raw multihash instead of raw CID as key in Head.AddProvider() (#156)
DHT clients expect messages to contain raw multihashes, not CIDs. This works for CIDv0 but not CIDv1, since a raw CIDv0 is a raw multihash but a raw CIDv1 is not.

I believe the impact of this is that hydras will not find providers for CIDv1's added via Head.AddProvider (I believe the provider prefetching uses this), since clients will be requesting a different key than what was used in the datastore.